refactor: refactor services/trainer
