+++ services/agents/app/core/health.py
@@ -4,7 +4,7 @@ from typing import List, Dict
 
 from app.logs import get_logger
 from app.config import config_base_llm
-from app.services.tasker import tasker
+from app.services.tasker import tasker_client
 from app.services.ml_models import ml_models_client
 from app.services.metrics import health as metrics_health
 from app.services.trainer import health as trainer_health
@@ -90,7 +90,7 @@ def check_health_all() -> Dict[str, List]:
     info["services"] = {
         "datasets": datasets_healt(),
         "ml_models": ml_models_client.health(),
-        "tasker": tasker.health(),
+        "tasker": tasker_client.health(),
         "trainer": trainer_health(),
         "metrics": metrics_health()
     }
